more node renames around debuggingManager, also commented out FileSystem usage in kernelDebugAdapter
