+++ src/platform/debugger/jupyter/debuggingManager.node.ts
@@ -20,7 +20,7 @@ import {
 import * as path from '../../../platform/vscode-path/path';
 import { IKernel, IKernelProvider } from '../../../kernels/types';
 import { IConfigurationService, IDisposable } from '../../common/types';
-import { KernelDebugAdapter } from './kernelDebugAdapter.node';
+import { KernelDebugAdapter } from './kernelDebugAdapter';
 import { IExtensionSingleActivationService } from '../../activation/types';
 import { ContextKey } from '../../common/contextKey';
 import { IApplicationShell, ICommandManager, IVSCodeNotebook } from '../../common/application/types';
@@ -33,9 +33,9 @@ import { DebuggingTelemetry, pythonKernelDebugAdapter } from '../constants';
 import { sendTelemetryEvent } from '../../../telemetry';
 import { DebugCellController, RunByLineController } from './debugControllers';
 import { assertIsDebugConfig, IpykernelCheckResult, isUsingIpykernel6OrLater } from './helper';
-import { Debugger } from './debugger.node';
+import { Debugger } from './debugger';
 import { INotebookControllerManager } from '../../../notebooks/types';
-import { IFileSystem } from '../../common/platform/types.node';
+//import { IFileSystem } from '../../common/platform/types.node';
 
 /**
  * The DebuggingManager maintains the mapping between notebook documents and debug sessions.
@@ -57,7 +57,7 @@ export class DebuggingManager implements IExtensionSingleActivationService, IDeb
         @inject(ICommandManager) private readonly commandManager: ICommandManager,
         @inject(IApplicationShell) private readonly appShell: IApplicationShell,
         @inject(IVSCodeNotebook) private readonly vscNotebook: IVSCodeNotebook,
-        @inject(IFileSystem) private fs: IFileSystem,
+        //@inject(IFileSystem) private fs: IFileSystem,
         @inject(IConfigurationService) private settings: IConfigurationService,
         @inject(IPlatformService) private platform: IPlatformService
     ) {
@@ -380,7 +380,8 @@ export class DebuggingManager implements IExtensionSingleActivationService, IDeb
                         session,
                         debug.document,
                         kernel.session,
-                        this.fs,
+                        // IANHU TOOD: Need to resolve
+                        //this.fs,
                         kernel,
                         this.platform
                     );
